ui/requirements.txt
streamlit>=1.44.0
httpx>=0.28.0
Pillow>=11.0.0
numpy>=2.2.0
opencv-python-headless>=4.11.0.86
soundfile>=0.13.0
python-dotenv>=1.1.0
pytest>=8.3.0
